rewoo-summarizer/requirements-rewoo.txt
transformers>=4.44.0
torch>=2.4.0
accelerate>=0.34.0
requests>=2.32.0
python-dotenv>=1.0.1
